+++ pytensor/tensor/rewriting/blockwise.py
@@ -8,13 +8,13 @@ from pytensor.graph.traversal import apply_ancestors
 from pytensor.tensor.basic import Alloc, ARange, alloc, shape_padleft
 from pytensor.tensor.blockwise import Blockwise, _squeeze_left
 from pytensor.tensor.math import Dot
+from pytensor.tensor.reshape import Reshape
 from pytensor.tensor.rewriting.basic import (
     register_canonicalize,
     register_specialize,
     register_stabilize,
 )
 from pytensor.tensor.rewriting.elemwise import InplaceGraphOptimizer
-from pytensor.tensor.shape import Reshape
 from pytensor.tensor.subtensor import (
     AdvancedIncSubtensor,
     AdvancedSubtensor,
