+++ tests/link/mlx/test_shape.py
@@ -4,7 +4,8 @@ import pytest
 import pytensor.tensor as pt
 from pytensor.compile.ops import DeepCopyOp, ViewOp
 from pytensor.configdefaults import config
-from pytensor.tensor.shape import Shape, Shape_i, reshape
+from pytensor.tensor.reshape import reshape
+from pytensor.tensor.shape import Shape, Shape_i
 from pytensor.tensor.type import iscalar, vector
 from tests.link.mlx.test_basic import compare_mlx_and_py
